% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseDir.R
\name{browseDirUI}
\alias{browseDirUI}
\title{Directory Selection Module UI}
\usage{
browseDirUI(id, label)
}
\arguments{
\item{id}{A namespace identifier for the module.}

\item{label}{label of the button. default "Set Output Directory"}
}
\value{
A UI definition for the module.
}
\description{
This module provides a UI for selecting a directory for output processing.
}
\examples{
if (interactive()) {
 ui <- fluidPage(
   mainPanel(
     browseDirUI("vcfDir1"),
     browseDirUI("vcfDir2")
   )
 )
 
 server <- function(input, output, session) {
 Work_Dir <- normalizePath("~")
   selected_dir1 <- browseDirServer("vcfDir1")
   selected_dir2 <- browseDirServer("vcfDir2")
   
   observe({
     print(paste("Directory 1:", selected_dir1()))
     print(paste("Directory 2:", selected_dir2()))
   })
}
 
 shinyApp(ui = ui, server = server)
}
}
