% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseFile.R
\name{browseFileServer}
\alias{browseFileServer}
\title{Browse File Server}
\usage{
browseFileServer(id, extension, output, session)
}
\arguments{
\item{id}{The input id for the module.}

\item{extension}{The file extension to filter by.}

\item{output}{output needed for server side}

\item{session}{session needed for server side}
}
\value{
A reactive value containing the file path.
}
\description{
This function is the server logic for the browse file module. 
It stores the selected file path in a global reactive value.
}
\examples{
\dontrun{
  library(shiny)

  ui <- fluidPage(
    browseFileUI("file_input_ui", extension = ".csv")
  )

  server <- function(input, output, session) {
    browseFileServer("file_input_server", extension = ".csv")
  }

  shinyApp(ui, server)
}
}
