% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseFile_server_side.R
\name{browseFileServer_server_side}
\alias{browseFileServer_server_side}
\title{Directory Selection Module Server}
\usage{
browseFileServer_server_side(id, filetype, workspace)
}
\arguments{
\item{id}{A namespace identifier for the module.}

\item{filetype}{A character sting specifying the files type to use for starting 
directory path in the reactiveValues object. default c('csv', 'tsv', 'txt')}

\item{workspace}{set default work directory}
}
\value{
None. This function is used for its side effects.
It stores the path of the selected directory to the provided reactiveValues object.
}
\description{
This module handles the server-side logic for selecting a directory for output processing.
}
\examples{
if (interactive()) {
 ui <- fluidPage(
   mainPanel(
     browseFileUI_server_side("vcfDir1"),
     browseFileUI_server_side("vcfDir2")
   )
 )
 
 server <- function(input, output, session) {
   Work_Dir <- normalizePath("~")
   selected_file1 <- browseFileServer_server_side("vcfDir1")
   selected_file2 <- browseFileServer_server_side("vcfDir2")
   
   observe({
     print(paste("Directory 1:", selected_file1()))
     print(paste("Directory 2:", selected_file2()))
   })
}
 
 shinyApp(ui = ui, server = server)
}
}
