% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseFile_server_side.R
\name{browseFileUI_server_side}
\alias{browseFileUI_server_side}
\title{Directory Selection Module UI}
\usage{
browseFileUI_server_side(id, label)
}
\arguments{
\item{id}{A namespace identifier for the module.}

\item{label}{label of the button. default "Set Output Directory"}
}
\value{
A UI definition for the module.
}
\description{
This module provides a UI for selecting a directory for output processing.
}
\examples{
if (interactive()) {
 ui <- fluidPage(
   mainPanel(
     browseFileUI_server_side("vcfDir1"),
     browseFileUI_server_side("vcfDir2")
   )
 )
 
 server <- function(input, output, session) {
 Work_Dir <- normalizePath("~")
   selected_file1 <- browseFileServer_server_side("vcfDir1")
   selected_file2 <- browseFileServer_server_side("vcfDir2")
   
   observe({
     print(paste("Directory 1:", selected_file1()))
     print(paste("Directory 2:", selected_file2()))
   })
}
 
 shinyApp(ui = ui, server = server)
}
}
