% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseFile.R
\name{browserFileUI}
\alias{browserFileUI}
\title{Browse File}
\usage{
browserFileUI(id, extension, label = paste0("Select ", extension, " File"))
}
\arguments{
\item{id}{The input id for the module.}

\item{extension}{The file extension to filter by.}

\item{label}{The label for the file input.}
}
\value{
A list containing the reactive value of the file path.
}
\description{
A function to browse a file with a specific extension.
}
\examples{
\dontrun{
  library(shiny)

  ui <- fluidPage(
    browserFileUI("file_input_ui", extension = ".csv")
  )

  server <- function(input, output, session) {
    browserFileServer("file_input_server", extension = ".csv")
  }

  shinyApp(ui, server)
}
}
