% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdownUI.R
\name{dropdownUI}
\alias{dropdownUI}
\title{Dropdown UI Function}
\usage{
dropdownUI(id, label, content)
}
\arguments{
\item{id}{The ID of the dropdown module.}

\item{label}{The label for the dropdown button.}

\item{content}{The content to be displayed in the dropdown menu.}
}
\value{
A Shiny UI element for the dropdown button.
}
\description{
Creates a dropdown button with a customizable content area.
}
\examples{
if (interactive()) {
library(shiny)
ui <- fluidPage(
  dropdownUI("dropdown1", "FastQ List", DT::dataTableOutput("fastq_list_DT"))
)
server <- function(input, output, session) {}
shinyApp(ui = ui, server = server)
}
}
