% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iconDialog.R
\name{iconDialogUI}
\alias{iconDialogUI}
\title{Icon Dialogue Module UI}
\usage{
iconDialogUI(id, icon)
}
\arguments{
\item{id}{The id of the module.}

\item{icon}{Set Icon model}
}
\description{
This module creates an info icon that displays a message when hovered over or clicked.
}
\examples{
\dontrun{
ui <- fluidPage(
  iconDialogUI("info_module", icon),
  # Add other UI elements as needed
)

server <- function(input, output, session) {
  iconDialogServer("info_module", title = "Custom Title", message = "Custom Message")
  # Add other server logic as needed
}

shinyApp(ui = ui, server = server)
}

}
