% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressBar.R
\name{progressBarServer}
\alias{progressBarServer}
\title{Progress Bar Module Server}
\usage{
progressBarServer(id, scriptPath, args_list, startMessage, endMessage,
buttonLabel, buttonVisibility, buttonStyle)
}
\arguments{
\item{id}{A unique identifier for the module.}

\item{scriptPath}{Path to the Bash script to be executed.}

\item{args_list}{A list of arguments to pass to the script.}

\item{startMessage}{Message displayed when the script starts.}

\item{endMessage}{Message displayed when the script completes.}

\item{buttonLabel}{A label for the action button.}

\item{buttonVisibility}{A reactive expression to control the visibility of the button.}

\item{buttonStyle}{Custom CSS styling for the action button.}
}
\value{
None. This function is used for its side effects.
}
\description{
Handles the server logic for the progress bar module.
}
\examples{
if (interactive()) {
# In your server definition
progressBarServer("progress1", scriptPath = "script1.sh",
args_list = c("-i", "input_dir", "-o", "output_dir"), 
startMessage = "Starting script 1...", endMessage = "Script 1 completed",
buttonLabel= "start script 1", buttonVisibility = reactive({ TRUE }))

progressBarServer("progress2", scriptPath = "script2.sh",
args_list = c("-i", "input_dir", "-o", "output_dir"),
startMessage = "Starting script 2...", endMessage = "Script 2 completed",
buttonLabel= "start scrip2t", buttonVisibility = reactive({ FALSE }),
buttonStyle = FALSE)
}
}
