% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressBar.R
\name{progressBarUI}
\alias{progressBarUI}
\title{Progress Bar Module UI}
\usage{
progressBarUI(id, buttonLabel = "Start Script")
}
\arguments{
\item{id}{A unique identifier for the module.}

\item{buttonLabel}{A label for the action button.}
}
\value{
A `tagList` containing the action button and the progress bar UI elements.
}
\description{
Creates the UI for the progress bar module with customizable button styling and visibility.
}
\examples{
if (interactive()) {
# In your UI definition
progressBarUI("progress1", buttonLabel = "Run Script 1",
buttonStyle = "position: relative; transform: translateY(-50\%);
left: 50\%; transform: translateX(-50\%); margin-top:30px")
progressBarUI("progress2", buttonLabel = "Run Script 2")
}
}
