% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.dragen.R
\name{shiny.dragen}
\alias{shiny.dragen}
\title{Launch shiny.dragen with default browser}
\usage{
shiny.dragen()
}
\value{
shiny webpage
}
\description{
Launch shiny.dragen with default browser
}
\examples{
ShinyApp <-  1
\dontrun{
shiny.dragen()
}

}
