% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseDir.R
\name{browseDirServer}
\alias{browseDirServer}
\title{Directory Selection Module Server}
\usage{
browseDirServer(id, rv, dir_key, filetype)
}
\arguments{
\item{id}{A namespace identifier for the module.}

\item{rv}{A reactiveValues object for storing the selected directory path.}

\item{dir_key}{A character string specifying the key to use for storing 
the directory path in the reactiveValues object. default: /home.}

\item{filetype}{A character sting specifying the files type to use for starting 
directory path in the reactiveValues object. default c('csv', 'tsv', 'txt')}
}
\value{
None. This function is used for its side effects.
It stores the path of the selected directory to the provided reactiveValues object.
}
\description{
This module handles the server-side logic for selecting a directory for output processing.
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinyFiles)

  ui <- fluidPage(
    mainPanel(
      browseDirUI("vcfDir1")
    )
  )

  server <- function(input, output, session) {
    rv <- reactiveValues()
    browseDirServer("vcfDir1", rv, "vcf_folder_path", "filetype")
  }

  shinyApp(ui = ui, server = server)
}
}
