% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browseDir.R
\name{browseDirUI}
\alias{browseDirUI}
\title{Directory Selection Module UI}
\usage{
browseDirUI(id, label)
}
\arguments{
\item{id}{A namespace identifier for the module.}

\item{label}{label of the button. default "Set Output Directory"}
}
\value{
A UI definition for the module.
}
\description{
This module provides a UI for selecting a directory for output processing.
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinyFiles)

  ui <- fluidPage(
    mainPanel(
      browseDirUI("vcfDir1")
    )
  )

  server <- function(input, output, session) {
    browseDirServer("vcfDir1", rv, "vcf_folder_path", "filetype")
  }

  shinyApp(ui = ui, server = server)
}
}
